Fixing "adb shell cmd statusbar expand-settings" command in split shade
The command was not working because it used `flingSettings` method in lockscreen and every motion in split shade in lockscreen is now managed by LockscreenShadeTransitionController.
Circumventing LockscreenShadeTransitionController means some visibilities and alpha were out of sync and broken. Now we're using that class directly for split shade case and everything works smoothly.

Fixes: 228796547
Test: manual + NotificationPanelViewControllerTest
Change-Id: Icfadaf01cb95bd3590d6e2db0841fec26e64d5b8