Separate timeNumberFormatTrivialFormatLong into its own file and delete HostnameVerifierTest.
The timeNumberFormatTrivialFormatLong test has been occasionally failing
since ICU has a lot of caching and interactions with Gc, which makes it
unstable. Separating it into it's own file should reduce noise levels.

In addition HostnameVerifierTest fails when presubmit devices lack
internet collections, and is not a very consistent benchmark due to DNS internals, so we will delete it.

Test: atest {TestName}
Bug: 232196040
Bug: 232196536
Change-Id: I70e3950f7914f7d32cab2d81713ac109d6fe641c