Partially undo "Try to avoid overlapping job execution."
This reverts the behavior change in
commit f4c4406cd0e29b4d0d39ec7ad25dd3a4a7ef242c.

However, we log the instances where jobs are started while a similar one
is already running. There can be legitimate instances when there's an
overlap, so no reason to log as a WTF. Limit logging to when DEBUG is
enabled to avoid a performance hit on end-user devices.

Some documentation and logging additions were retained.

Reason for revert: Not necessary.

Bug: 231148615
Test: N/A
Change-Id: I5a89ad3dbd5c5e94b262485e8a418086ca862e9b