Revert "Rebind ContentCaptureService when the binderDied."
This reverts commit 60f16a9fc60ca229f4889d2d75c93cf3333b3ced.

We introduce a rebind mechanism to rebind died ContentCaptureService
because the apps tht has dependency with ContentCaptureService. But
it looks like the rate is raised for app to get the null content
capture manager.

Because the underlying issue is fixed, we don't need this workaround
now. We can also monitor if the null content capture sympton is
resolved.

Bug: 227353916
Test: manual. Add dependency and update the app, the ContentCapture
can be rebind without the this fix.
Reason for revert: To monitor if the null content capture issue
caused by this change.

Change-Id: Ibe5d6e2b5f1434243b655289fdc60174ff33a5c0