Clear ActivityRecord#mDismissKeyguard flag after app launches.
This flag is used by the SysUI to dismiss an insecure keyguard
when an app is launched from an insecure lock screen. The flag
should only affect the app launch once, so needs to be cleared
once the flag is processed.

Bug: 234583880
Test: atest KeyguardTests#testDismissKeyguard_fromActivityOption_onlyOnce
Change-Id: Ia4179b40de7d45b2199b175a6dcc24016dfba091
(cherry picked from commit e8dc98c80cbee2049733fdf61331764a47eca44b)