Make IInputMethodSession async even for in-proc cases
This CL applies the same approach I used for IInputMethodClient
recently [1], that is, to use a dedicated worker thread to invoke
binder IPCs to achieve the async semantics even if IInputMethodSession
is not a BinderProxy, which happens when both the IME and IME client
are running in the same process.

When the IME and IME client are running in different processes, this
CL literally changes nothing.

 [1]: I0a7dd5162c4b94599b896420b8427913fa772414
      4d849adbef14d7455fd507875667bf88c2fa8302

Fix: 236147652
Test: atest CtsInputMethodTestCases:InProcessImeTest
Change-Id: I56f3198e54f689adc9541f2c2644ee19bcc6deed