Remove unused InputMethodManagerInternal#isStylusHandwritingAvailable()
This CL removes an unused internal method

  InputMethodManagerInternal#isStylusHandwritingAvailable(),

which was added recently [1].

Currently we are discussing multiple projects that may affect the
number of active bindings that InputMethodManagerService manages.
Depending on how such discussions go, we may increase or decrease the
number of active IME sessions, and the concent of

  "isStylusHandwritingAvailable"

may need to be multi-user aware for callers within system_server.
Thus it may make more sense to not expose

  InputMethodManagerInternal#isStylusHandwritingAvailable()

until someone actually start using it in system_server.

 [1]: I0c606e94780ae6636585ab36b751fbf6e90a1cbf
      e14a89d32832ab06ebe05570e24c1dd043d1da7a

Bug: 228514000
Test: presubmit
Change-Id: Ia6471a097715bb25fdb0a5a1cb82d2c77d4c3024