Set mAlreadyRequestedSync only when in a BLASTSync
WindowState#useBLASTSync returns true if in a blast sync but also if
there are mDrawHandlers. In that case, we only want to mark
mAlreadyRequestedSync to true if we know it's a blast sync since the
unmark is done when the sync is complete.

Test: Pip expand no delay
Fixes: 236882306
Change-Id: Ifb5785b93888a0fb3e27180005ea1b3940a293b7
(cherry picked from commit 6f8c25498cca57f084cb659f03c5c73ef8d8d516)
Merged-In: Ifb5785b93888a0fb3e27180005ea1b3940a293b7