Centralize upper/lower time bound logic
Move the upper / lower bound constants for valid time suggestions into a
single class TimeDetectorHelper, which can be accessed by both server
and client (SettingsUI) code. This reduces the number of places that
need to be touched when we change those values.

See also the associated change in packages/apps/Settings which removes
bounds checks from the client code and switches that code over to use
TimeDetectorHelper methods.

Test: services/tests/servicestests/src/com/android/server/timedetector/
Bug: 228967927
Change-Id: I952ecd4fc2e0d80ffeda4e288fb838253142953e