Remove PackageInfo from ApexPackageInfo
Consumers inside system_server should use AndroidPackage/PackageState
directly.

This may regress some metrics as it pushes PackageInfo generation to
call time instead of cached ahead of time, but this method is more
accurate as it allows filling any user-specific fields and mirrors
how _Info objects are generated in the rest of the system.

Bug: 239965524

Test: atest ApexManagerTest

Change-Id: Ibdea719c75a9c72ac144a1f30692d04d280c6309