Remove redundant enforceCallingPermission()
This CL aims to simplify security checks for

  IInputMethodManager#onShellCommand().

Basically the calling UID check [1] remains to be the main security
policy for that IPC method.  Context#enforceCallingPermission() checks
added in later CLs [2][3] are just redundant with an assumption that
the shell package has those permissions.

For simplicity and readability, let's keep only the most relevant
security check for IMMS#onShellCommand() method.

 [1]: If87189563ccaacd4f9c666bab4f9ad08a9343084
      26869bd26c2591932ac6ade9f1b5f330eecfeb1c
 [2]: I6fd47b5cc1e7da7222774df20247a2c69a70f45b
      daec8f6ad88437613e82f18c7d9d802fff89579f
 [3]: If91137e0c144d90ac3046f30607e36bb957a7e0e
      32f7d21da6f41a9b9619b93772de969c8274966c

Bug: 34886274
Test: atest CtsInputMethodServiceHostTestCases:ShellCommandFromAppTest
Change-Id: I19a6833bb7ff6100d4e9428243005377a7fbf432